requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
cachetools==4.2.4
certifi==2023.5.7
charset-normalizer==3.2.0
dj-database-url==0.5.0
Django
django-filter==23.2
django-googledrive-storage==1.6.0
djangorestframework
google-api-core==2.11.1
google-api-python-client==2.93.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.1
gunicorn==20.1.0
httplib2==0.22.0
idna==3.4
importlib-metadata==6.8.0
Markdown==3.4.3
Pillow
pre-commit
protobuf==4.23.4
psycopg2-binary
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
rsa==4.9
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.3
validate-email==1.3
whitenoise==5.3.0
zipp==3.16.2
